feat: Support Honkai Star Rail pre-download package
